requirements.txt
rpi.bme280
pandas
Flask
